New implementation of ALSoundOut: - Passing SoundOut-Unittests on Windows - Context-Switching on Linux(and Mac?)

+++ CSCore/SoundOut/AL/ALInterops.cs
@@ -1,7 +1,7 @@
-
 using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
+// ReSharper disable InconsistentNaming
 
 namespace CSCore.SoundOut.AL
 {
@@ -87,6 +87,8 @@ namespace CSCore.SoundOut.AL
         internal static extern void alGetSource3f(uint sourceId, ALSourceParameters param, out float val1,
             out float val2, out float val3);
 
+
+
         [DllImport("OpenAL32.dll", CallingConvention = CallingConvention.Cdecl)]
         internal static extern void alSourcef(uint sourceId, ALSourceParameters param, float value);
 
@@ -98,7 +100,9 @@ namespace CSCore.SoundOut.AL
             float val3);
 
         [DllImport("OpenAL32.dll", CallingConvention = CallingConvention.Cdecl)]
-        internal static extern void alSourcei(uint sourceId, ALSourceParameters param, float val1);
+        internal static extern void alSourcei(uint sourceId, ALSourceParameters param, int val1);
+
+
 
         [DllImport("OpenAL32.dll", CallingConvention = CallingConvention.Cdecl)]
         internal static extern void alGenBuffers(int count, uint[] bufferIDs);
